python-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
transformers==4.45.2
torch>=2.3.0
Pillow==10.4.0
python-multipart==0.0.12
accelerate==1.0.1

